data_reader/requirements.txt
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
docutils==0.20.1
Flask==3.0.2
Flask-Login==0.6.3
fonttools==4.50.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==5.1.0
jaraco.functools==4.0.0
jeepney==0.8.0
Jinja2==3.1.3
keyring==25.1.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplot==0.1.9
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.2.0
nh3==0.2.17
numpy==1.26.4
packaging==24.0
paho-mqtt==2.1.0
pandas==2.2.1
pillow==10.3.0
pkginfo==1.10.0
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pycparser==2.22
Pygments==2.17.2
pyloco==0.0.139
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
readme_renderer==43.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
SecretStorage==3.3.3
SimpleWebSocketServer==0.1.2
six==1.16.0
twine==5.0.0
typing==3.7.4.3
tzdata==2024.1
urllib3==2.2.1
ushlex==0.99.1
websocket-client==1.7.0
Werkzeug==3.0.2
zipp==3.18.1
